Memoized Solution
Solution that goes sums multiples of 3 and then 5.
Uses a set to keep track of numbers it already counted so it doesn't
count numbers like 15.
This solution is slower than the simple iterative solution. This is due
to the extra checks and the use of the set, which also takes up much
more memory than the simple iterative solution.